time_series/PJ/requirements.txt
aiogram==3.4.1
yfinance==1.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
statsmodels==0.14.2
matplotlib==3.8.4
scipy==1.11.4
tensorflow==2.20.0
python-dotenv==1.0.1
